% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recall.R
\name{map_recall}
\alias{map_recall}
\title{Function for getting lat and long data from addresses}
\usage{
map_recall(data)
}
\arguments{
\item{data}{The dataframe used as input with addresses to map latitude and longitudinally}
}
\description{
Function for getting lat and long data from addresses
}
