% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_date.R
\name{recall_date}
\alias{recall_date}
\title{This function scrapes the openFDA API for food product recall enforcement reports based on user inputs of date and time}
\usage{
recall_date(
  api_key,
  center_classification_date = NULL,
  limit = NULL,
  product_description = NULL,
  recall_initiation_date = NULL,
  recalling_firm = NULL,
  report_date = NULL,
  search_mode = NULL,
  status = NULL,
  termination_date = NULL
)
}
\arguments{
\item{api_key}{Your free api key from FDA website}

\item{center_classification_date}{A way the FDA classifies a date}

\item{limit}{The number of rows to return for that query}

\item{product_description}{Description of product}

\item{recall_initiation_date}{Date for which recall was initiated}

\item{recalling_firm}{The company recalling the product}

\item{report_date}{The date the FDA issued the enforcement report for the product recall}

\item{search_mode}{This gives the user flexibility to search for exact matches of inputs or any combination of inputs}

\item{status}{The status of the recall}

\item{termination_date}{The date the recall was terminated}
}
\value{
A data frame with the returned results of the users query to the API
}
\description{
This function scrapes the openFDA API for food product recall enforcement reports based on user inputs of date and time
}
\examples{
\dontrun{
recall_date(api_key = api_key, termination_date = "2021 to 2022)
recall_date(api_key = api_key, report_date = "January 1, 2023, status = "Ongoing")
recall_date(api_key = api_key, product_description = "Milk", recall_initiation_date = "2022")
}
}
