% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_location.R
\name{recall_location}
\alias{recall_location}
\title{This function scrapes the openFDA API for food product recall enforcement reports based on user inputs of location}
\usage{
recall_location(
  api_key,
  limit = NULL,
  city = NULL,
  country = NULL,
  distribution_pattern = NULL,
  recalling_firm = NULL,
  search_mode = NULL,
  state = NULL,
  status = NULL
)
}
\arguments{
\item{api_key}{Your free api key from openFDA API website}

\item{limit}{The number of rows to return for that query}

\item{city}{City where food company is located}

\item{country}{The country where the food was produced}

\item{distribution_pattern}{Locations where food was distributed to}

\item{recalling_firm}{The company recalling the product}

\item{search_mode}{This gives the user flexibility to search for exact matches of inputs or any combination of inputs}

\item{state}{The U.S. state in which the recalling firm is located}

\item{status}{The status of the recall}
}
\value{
A data frame with the returned results of the users query to the API
}
\description{
This function scrapes the openFDA API for food product recall enforcement reports based on user inputs of location
}
\examples{
\dontrun{
recall_location(api_key = api_key, city = "Ames", state = "Iowa")
recall_location(api_key = api_key, recalling_firm = "Target")
recall_location(api_key = api_key, distribution_pattern = "Colorado", status = "Ongoing")
recall_location(api_key = api_key, city = "Iowa City, Ames, Des Moines", state = "Iowa")
}
}
